feat: reindex elasticsearch when running migrations
