+++ packages/migration/run-migrations.sh
@@ -37,3 +37,6 @@ yarn --cwd $SCRIPT_PATH migrate-mongo status --file $USER_MGNT_CONFIG
 # performance migration
 yarn --cwd $SCRIPT_PATH migrate-mongo up --file $PERFORMANCE_CONFIG
 yarn --cwd $SCRIPT_PATH migrate-mongo status --file $PERFORMANCE_CONFIG
+
+# search migration / reindex
+yarn --cwd $SCRIPT_PATH reindex-search
